% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{add_labels}
\alias{add_labels}
\title{Function for adding a row with treatment labels for the figures.}
\usage{
add_labels(amce, labels = wiggle::treat_labs())
}
\arguments{
\item{amce}{Estimates from \code{amce()}.}

\item{labels}{The dataframe containing the labels. It expects a
column \code{treatment} corresponding to the treatment variables
and a column \code{treatment_label} with their labels. Defaults
to \code{wiggle::treat_labs()}, which is included in the
compendium.}
}
\description{
Function for adding a row with treatment labels for the figures.
}
